Separate Ben's lifecycle refactors from the rest of the commit
